Move validationHelper to helpers folder
